// src/app/module/auth/@shared/interfaces/auth.interface.ts
export interface Rules {
  data: string
  rule: string
}

export interface Form {
  registration: string
  name: string
  email: string
  password: string
  repeatPassword: string
  rememberData: string
  formDescription: string
  reg: string
  login: string
  loginAccount: string
  loginDescription: string
}